+++ resolwe_bio/tests/workflows/test_rna_seq.py
@@ -201,8 +201,7 @@ class RNASeqWorkflowTestCase(KBBioProcessTestCase):
         self.assertFields(fc_paired, 'species', 'Homo sapiens')
 
     @with_resolwe_host
-    @tag_process('workflow-bbduk-star-featurecounts-single', 'workflow-bbduk-star-featurecounts-paired',
-                 'workflow-bbduk-star-featurecounts-qc-single', 'workflow-bbduk-star-featurecounts-qc-paired')
+    @tag_process('workflow-bbduk-star-featurecounts-qc-single', 'workflow-bbduk-star-featurecounts-qc-paired')
     def test_bbduk_star_featurecounts_workflow(self):
         with self.preparation_stage():
             reads = self.prepare_reads(['hs sim_reads_single.fastq.gz'])
@@ -264,36 +263,6 @@ class RNASeqWorkflowTestCase(KBBioProcessTestCase):
             'quantification': {
                 'annotation': annotation.id,
             },
-        }
-        self.run_process('workflow-bbduk-star-featurecounts-single', inputs)
-        for data in Data.objects.all():
-            self.assertStatus(data, Data.STATUS_DONE)
-        feature_counts = Data.objects.get(process__slug='feature_counts')
-        self.assertFile(feature_counts, 'rc', 'feature_counts_rc_single.tab.gz', compression='gzip')
-        multiqc = Data.objects.get(process__slug='multiqc')
-        self.assertFileExists(multiqc, 'report')
-
-        inputs['preprocessing']['reads'] = paired_reads.id
-        self.run_process('workflow-bbduk-star-featurecounts-paired', inputs)
-        for data in Data.objects.all():
-            self.assertStatus(data, Data.STATUS_DONE)
-        feature_counts = Data.objects.filter(process__slug='feature_counts').last()
-        self.assertFile(feature_counts, 'rc', 'feature_counts_rc_paired.tab.gz', compression='gzip')
-        multiqc = Data.objects.filter(process__slug='multiqc').last()
-        self.assertFileExists(multiqc, 'report')
-
-        inputs = {
-            'preprocessing': {
-                'reads': reads.id,
-                'adapters': [adapters.id],
-                'custom_adapter_sequences': ['ACTGACTGACTG', 'AAACCCTTT'],
-            },
-            'alignment': {
-                'genome': star_index.id,
-            },
-            'quantification': {
-                'annotation': annotation.id,
-            },
             'qc': {
                 'rrna_reference': rrna_star_index.id,
                 'globin_reference': globin_star_index.id,
@@ -320,6 +289,15 @@ class RNASeqWorkflowTestCase(KBBioProcessTestCase):
         self.assertFields(globin, 'build', 'globin')
         multiqc = Data.objects.filter(process__slug='multiqc').last()
         self.assertFileExists(multiqc, 'report')
+
+        # test the pipeline without the adapter sequences specified
+        del inputs['preprocessing']['adapters']
+        del inputs['preprocessing']['custom_adapter_sequences']
+        self.run_process('workflow-bbduk-star-featurecounts-qc-paired', inputs)
+        for data in Data.objects.all():
+            self.assertStatus(data, Data.STATUS_DONE)
+        feature_counts = Data.objects.filter(process__slug='feature_counts').last()
+        self.assertFile(feature_counts, 'rc', 'feature_counts_rc_paired_wo_adapter_trim.tab.gz', compression='gzip')
 
     @with_resolwe_host
     @tag_process('workflow-custom-cutadapt-star-htseq-single', 'workflow-custom-cutadapt-star-htseq-paired')
@@ -514,63 +492,3 @@ class RNASeqWorkflowTestCase(KBBioProcessTestCase):
         self.assertFile(workflow, 'rc', 'workflow_rnaseq_paired_rc.tab.gz', compression='gzip')
         self.assertFields(workflow, 'exp_type', 'TPM')
         self.assertFields(workflow, 'source', 'DICTYBASE')
-
-    @with_resolwe_host
-    @tag_process('workflow-cutadapt-star-featurecounts-single', 'workflow-cutadapt-star-featurecounts-paired')
-    def test_cutadapt_star_featurecounts_workflow(self):
-        with self.preparation_stage():
-            reads = self.prepare_reads(['SRR7455803_1_short.fastq.gz'])
-            paired_reads = self.prepare_paired_reads(
-                mate1=['SRR7455803_1_short.fastq.gz'],
-                mate2=['SRR7455803_2_short.fastq.gz'])
-
-            inputs = {
-                'src': 'PGSC_ITAG_test.gff',
-                'build': 'stu_dm_itag-pgsc_4.04_nib',
-                'source': 'PGSC/ITAG',
-                'species': 'Solanum tuberosum',
-            }
-            gff3 = self.run_process('upload-gff3', inputs)
-            star_index_fasta = self.prepare_adapters(fn='StPGSC4.04n_Chr09.fasta.gz')
-            inputs = {'annotation': gff3.id, 'genome2': star_index_fasta.id}
-            star_index = self.run_process('alignment-star-index', inputs)
-
-        for data in Data.objects.all():
-            self.assertStatus(data, Data.STATUS_DONE)
-
-        self.run_process(
-            'workflow-cutadapt-star-featurecounts-single', {
-                'preprocessing': {
-                    'reads': reads.id,
-                },
-                'alignment': {
-                    'genome': star_index.id,
-                },
-                'quantification': {
-                    'annotation': gff3.id
-                }
-            }
-        )
-        for data in Data.objects.all():
-            self.assertStatus(data, Data.STATUS_DONE)
-
-        featurecounts = Data.objects.last()
-        self.assertFile(featurecounts, 'rc', 'workflow_cutadapt_star_fc_single.tab.gz', compression='gzip')
-
-        self.run_process(
-            'workflow-cutadapt-star-featurecounts-paired', {
-                'preprocessing': {
-                    'reads': paired_reads.id,
-                },
-                'alignment': {
-                    'genome': star_index.id,
-                },
-                'quantification': {
-                    'annotation': gff3.id
-                }
-            }
-        )
-        for data in Data.objects.all():
-            self.assertStatus(data, Data.STATUS_DONE)
-        featurecounts = Data.objects.last()
-        self.assertFile(featurecounts, 'rc', 'workflow_cutadapt_star_fc_paired.tab.gz', compression='gzip')